# api/server/public/openApi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/app/wardrobe/profile": {
      "get": {
        "operationId": "WardrobeController_getProfile",
        "summary": "获取身材档案和偏好",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "衣橱管家APP"
        ]
      },
      "post": {
        "operationId": "WardrobeController_upsertProfile",
        "summary": "创建或更新身材档案和偏好",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertProfileDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "衣橱管家APP"
        ]
      }
    },
    "/app/wardrobe/clothing": {
      "get": {
        "operationId": "WardrobeController_listClothing",
        "summary": "查询衣橱单品",
        "parameters": [
          {
            "name": "pageNum",
            "required": true,
            "in": "query",
            "description": "当前分页",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "description": "每页数量",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "orderByColumn",
            "required": false,
            "in": "query",
            "description": "排序字段",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isAsc",
            "required": false,
            "in": "query",
            "description": "排序规则",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "衣橱管家APP"
        ]
      },
      "post": {
        "operationId": "WardrobeController_createClothing",
        "summary": "新增衣物",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateClothingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "衣橱管家APP"
        ]
      }
    },
    "/app/wardrobe/clothing/{id}": {
      "patch": {
        "operationId": "WardrobeController_updateClothing",
        "summary": "更新衣物",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "衣橱管家APP"
        ]
      },
      "delete": {
        "operationId": "WardrobeController_deleteClothing",
        "summary": "删除衣物",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "衣橱管家APP"
        ]
      }
    },
    "/app/wardrobe/ai/analyze-clothing": {
      "post": {
        "operationId": "WardrobeController_analyzeClothing",
        "summary": "AI识别衣物图片",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeClothingDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "衣橱管家APP"
        ]
      }
    },
    "/app/wardrobe/outfits/generate": {
      "post": {
        "operationId": "WardrobeController_generateOutfits",
        "summary": "AI生成搭配",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateOutfitDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "衣橱管家APP"
        ]
      }
    },
    "/app/wardrobe/outfits/history": {
      "get": {
        "operationId": "WardrobeController_history",
        "summary": "搭配历史",
        "parameters": [
          {
            "name": "pageNum",
            "required": true,
            "in": "query",
            "description": "当前分页",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "description": "每页数量",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "orderByColumn",
            "required": false,
            "in": "query",
            "description": "排序字段",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isAsc",
            "required": false,
            "in": "query",
            "description": "排序规则",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "衣橱管家APP"
        ]
      }
    },
    "/app/wardrobe/try-on/history": {
      "get": {
        "operationId": "WardrobeController_listTryOnHistory",
        "summary": "真人试穿历史",
        "parameters": [
          {
            "name": "pageNum",
            "required": true,
            "in": "query",
            "description": "当前分页",
            "schema": {
              "default": 1,
              "type": "number"
            }
          },
          {
            "name": "pageSize",
            "required": true,
            "in": "query",
            "description": "每页数量",
            "schema": {
              "default": 10,
              "type": "number"
            }
          },
          {
            "name": "orderByColumn",
            "required": false,
            "in": "query",
            "description": "排序字段",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "isAsc",
            "required": false,
            "in": "query",
            "description": "排序规则",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "衣橱管家APP"
        ]
      }
    },
    "/app/wardrobe/try-on/history/{id}": {
      "get": {
        "operationId": "WardrobeController_getTryOnHistory",
        "summary": "真人试穿历史详情",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "衣橱管家APP"
        ]
      },
      "delete": {
        "operationId": "WardrobeController_deleteTryOnHistory",
        "summary": "删除真人试穿历史",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "衣橱管家APP"
        ]
      }
    },
    "/app/wardrobe/outfits/{id}": {
      "get": {
        "operationId": "WardrobeController_getOutfit",
        "summary": "搭配详情",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "衣橱管家APP"
        ]
      },
      "delete": {
        "operationId": "WardrobeController_deleteOutfit",
        "summary": "删除搭配记录",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "衣橱管家APP"
        ]
      }
    },
    "/app/wardrobe/outfits/{id}/save": {
      "patch": {
        "operationId": "WardrobeController_saveOutfit",
        "summary": "收藏或取消收藏搭配",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveOutfitDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "衣橱管家APP"
        ]
      }
    },
    "/app/wardrobe/outfits/{id}/render": {
      "post": {
        "operationId": "WardrobeController_renderOutfit",
        "summary": "重新生成搭配平铺效果图",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "衣橱管家APP"
        ]
      }
    },
    "/app/wardrobe/outfits/{id}/try-on": {
      "post": {
        "operationId": "WardrobeController_startTryOn",
        "summary": "提交真人试穿任务",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartTryOnDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "衣橱管家APP"
        ]
      },
      "get": {
        "operationId": "WardrobeController_getTryOnStatus",
        "summary": "查询真人试穿任务状态",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "衣橱管家APP"
        ]
      }
    },
...
    "/app/wardrobe/challenge/score": {
      "post": {
        "operationId": "WardrobeController_scoreChallenge",
        "summary": "AI挑战评分",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScoreChallengeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "衣橱管家APP"
        ]
      }
    },
    "/app/wardrobe/submissions": {
      "post": {
        "operationId": "WardrobeController_createSubmission",
        "summary": "提交社区穿搭审核",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubmissionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "衣橱管家APP"
        ]
      }
    },
    "/app/wardrobe/feedback": {
      "post": {
        "operationId": "WardrobeController_createFeedback",
        "summary": "提交用户反馈",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateFeedbackDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "衣橱管家APP"
        ]
      }
    },
...
  },
  "info": {
    "title": "Nest-Admin",
    "description": "Nest-Admin 接口文档",
    "version": "2.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "components": {
    "securitySchemes": {
      "Authorization": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header",
        "name": "Authorization",
        "description": "请在请求头中携带 JWT 令牌，格式：Bearer <token>"
      }
    },
    "schemas": {
...
      "UpsertProfileDto": {
        "type": "object",
        "properties": {}
      },
      "CreateClothingDto": {
        "type": "object",
        "properties": {}
      },
      "AnalyzeClothingDto": {
        "type": "object",
        "properties": {}
      },
      "GenerateOutfitDto": {
        "type": "object",
        "properties": {}
      },
      "SaveOutfitDto": {
        "type": "object",
        "properties": {}
      },
      "StartTryOnDto": {
        "type": "object",
        "properties": {}
      },
      "ScoreChallengeDto": {
        "type": "object",
        "properties": {}
      },
      "CreateSubmissionDto": {
        "type": "object",
        "properties": {}
      },
      "CreateFeedbackDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}